@@ -15,8 +15,8 @@
  */
 package com.github.megatronking.netbare.net;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.github.megatronking.netbare.ip.Protocol;
 
