@@ -15,8 +15,8 @@
  */
 package com.github.megatronking.netbare.http;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.text.TextUtils;
 
 import java.text.DateFormat;
