@@ -16,8 +16,8 @@
 package com.github.megatronking.netbare.http;
 
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.github.megatronking.netbare.utils.CaseInsensitiveLinkedMap;
 import com.github.megatronking.netbare.stream.Stream;
