@@ -16,7 +16,7 @@
 package com.github.megatronking.netbare.http;
 
 import android.net.Uri;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.github.megatronking.netbare.NetBareUtils;
 
